feat(vim): replace netrw with nvim-tree for file exploration
This commit replaces the netrw plugin with the nvim-tree plugin for
file exploration in Neovim. The nvim-tree plugin is configured to
disable netrw, use floating window for file view, and custom glyphs
for git untracked files. The keybinding for file exploration is also
updated to use nvim-tree instead of netrw.